% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clade_phyloglm.R
\name{clade_phyglm}
\alias{clade_phyglm}
\title{Influential clade detection - Phylogenetic Logistic Regression}
\usage{
clade_phyglm(formula, data, phy, btol = 50, track = TRUE, clade.col,
  n.species = 5, ...)
}
\arguments{
\item{formula}{The model formula}

\item{data}{Data frame containing species traits with row names matching tips
in \code{phy}.}

\item{phy}{A phylogeny (class 'phylo') matching \code{data}.}

\item{btol}{Bound on searching space. For details see \code{phyloglm}}

\item{track}{Print a report tracking function progress (default = TRUE)}

\item{clade.col}{The name of a column in the provided data frame with clades
specification (a character vector with clade names).}

\item{n.species}{Minimum number of species in the clade in order to include
this clade in the leave-one-out deletion analyis. Default is \code{5}.}

\item{...}{Further arguments to be passed to \code{phyloglm}}
}
\value{
The function \code{clade_phyglm} returns a list with the following
components:

\code{formula}: The formula

\code{full.model.estimates}: Coefficients, aic and the optimised
value of the phylogenetic parameter (e.g. \code{alpha}) for the full model
without deleted species.

\code{clade.model.estimates}: A data frame with all simulation
estimates. Each row represents a deleted clade. Reported are the calculated
regression intercept (\code{intercept}), difference between simulation
intercept and full model intercept (\code{DFintercept}), the percentage change
 in intercept compared to the full model (\code{intercept.perc}) and intercept
 p-value (\code{pval.intercept}). All of these are also reported for the regression
slope (\code{DFslope} etc.). Additonally, model aic value (\code{AIC}) and
the optimised value (\code{optpar}) of the phylogenetic parameter are
reported.

\code{data}: Original full dataset.

\code{errors}: Clades where deletion resulted in errors.
}
\description{
Estimate the impact on model estimates of phylogenetic logistic regression after
removing clades from the analysis.
}
\details{
Currently only logistic regression using the "logistic_MPLE"-method from
\code{phyloglm} is implemented.

\code{clade_phyglm} detects influential clades based on
difference in intercept and/or slope when removing a given clade compared
to the full model including all species.

Currently, this function can only implement simple linear models (i.e.
\eqn{y = a + bx}). In the future we will implement more complex models.

Output can be visualised using \code{clade_plot}.
}
\author{
Gustavo Paterno & Gijsbert Werner
}
\references{
Here still: reference to phylolm paper + our own?
}
\seealso{
\code{\link[sensiPhy]{clade_phylm}}, \code{\link{influ_phyglm}},
\code{\link{sensi_plot}}
}

