% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/influ_gls.R
\name{influ_gls}
\alias{influ_gls}
\title{Leave-one-out-deletion analysis for gls phylogenetic regression.}
\usage{
influ_gls(formula, data, phy)
}
\arguments{
\item{formula}{A model formula}

\item{data}{A data frame containing variables that can be attributed to the taxa
at the tips of a phylogeny}

\item{phy}{A phylogeny (class 'phylo') to be matched to the data above}
}
\value{
The function \code{influ_gls} returns a list with the following
  components:

\code{errors} Species that showed erros during pgls fit

\code{formula} The model formula

\code{model_estimates} Full model estimates

\code{beta95_IC} Full model beta 95 confidence interval

\code{influential_species} Most influential species for beta

\code{results} A data frame with all simulation estimates. DFbeta and
  DFintercept represent absolute difference between full model and simulation.
}
\description{
\code{influ_gls} performs leave-one-out-deletion analysis for
\code{\link{gls}} phylogenetic linear regression. It removes one species at a time, fits
a gls model without the species and store the results of the model
estimates. The function repeats this procedure for all species in the dataset
and store all the results in a data.frame.
}
\details{
This functions only works for simple linear regression \eqn{y = bx +
  a}. Future implementation will deal with more complex models. This functions only
  works with Pagel correlation (see \code{corPagel}). If any error occur during simulation,
  check 'output$errors' for details.
}
\section{Warning}{
 This code is note fully checked. Please be aware.
}
\examples{
library(caper);library(ggplot2);library(gridExtra)
data(shorebird)
comp.data <- comparative.data(shorebird.tree, shorebird.data, Species, vcv=TRUE, vcv.dim=3)
# First we need to match tip.labels with rownames in data:
sp.ord <- match(shorebird.tree$tip.label, rownames(shorebird.data))
shorebird.data <- shorebird.data[sp.ord,]
# Now we can run the function influ_gls:
influ <- influ_gls(log(Egg.Mass) ~ log(M.Mass),data=shorebird.data,phy=shorebird.tree)
# Estimated parameters:
head(influ$results)
# Most influential species:
influ[[5]]
# Check for species with erros erros:
influ$errors
}
\seealso{
\code{\link{pgls}}, \code{\link{samp_pgls}}
}

