% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/influ_phyloglm.R
\name{influ_phyloglm}
\alias{influ_phyloglm}
\title{Leave-one-out-deletion analysis for gls phylogenetic regression. (GW: Make helpfile still!)
library(caper);library(ggplot2);library(gridExtra);library(phylolm)
data(shorebird)
sp.ord <- match(shorebird.tree$tip.label, rownames(shorebird.data))
shorebird.data <- shorebird.data[sp.ord,]
#Create a binary variable (large egg / small egg), for illustration purposes.
mean(shorebird.data$Egg.Mass)
shorebird.data$Egg.Mass.binary<-ifelse(shorebird.data$Egg.Mass>30,1,0) #Turn egg mass into a binary variable
table(shorebird.data$Egg.Mass.binary) #Mostly small eggs.
# Now we can run the function influ_phyloglm:
influ_phyloglm<-influ_phyloglm(formula = Egg.Mass.binary~M.Mass,data = shorebird.data,phy=shorebird.tree,btol = 50)
# Estimated parameters:
head(influ_phyloglm$results)
# Most influential species:
influ_phyloglm$influential_species
# Check for species with errors:
influ_phyloglm$errors}
\usage{
influ_phyloglm(formula, data, phy, btol = 50)
}
\description{
Leave-one-out-deletion analysis for gls phylogenetic regression. (GW: Make helpfile still!)
library(caper);library(ggplot2);library(gridExtra);library(phylolm)
data(shorebird)
sp.ord <- match(shorebird.tree$tip.label, rownames(shorebird.data))
shorebird.data <- shorebird.data[sp.ord,]
#Create a binary variable (large egg / small egg), for illustration purposes.
mean(shorebird.data$Egg.Mass)
shorebird.data$Egg.Mass.binary<-ifelse(shorebird.data$Egg.Mass>30,1,0) #Turn egg mass into a binary variable
table(shorebird.data$Egg.Mass.binary) #Mostly small eggs.
# Now we can run the function influ_phyloglm:
influ_phyloglm<-influ_phyloglm(formula = Egg.Mass.binary~M.Mass,data = shorebird.data,phy=shorebird.tree,btol = 50)
# Estimated parameters:
head(influ_phyloglm$results)
# Most influential species:
influ_phyloglm$influential_species
# Check for species with errors:
influ_phyloglm$errors
}

