% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/influ_phylolm.R
\name{influ_phylolm}
\alias{influ_phylolm}
\title{Influential species detection - Phylogenetic Linear Regression}
\usage{
influ_phylolm(formula, data, phy, model = "lambda", cutoff = 2,
  track = TRUE, ...)
}
\arguments{
\item{formula}{The model formula}

\item{data}{Data frame containing species traits with row names matching tips
in \code{phy}.}

\item{phy}{A phylogeny (class 'phylo') matching \code{data}.}

\item{model}{The phylogenetic model to use (see Details). Default is \code{lambda}.}

\item{cutoff}{The cutoff value used to identify for influential species
(see Details)}

\item{track}{Print a report tracking function progress (default = TRUE)}

\item{...}{Further arguments to be passed to \code{phylolm}}
}
\value{
The function \code{influ_phylolm} returns a list with the following
components:

\code{cutoff}: The value selected for \code{cutoff}

\code{formula}: The formula

\code{full.model.estimates}: Coefficients, aic and the optimised
value of the phylogenetic parameter (e.g. \code{lambda}) for the full model
without deleted species.

\code{influential_species}: List of influential species, both
based on standardised difference in interecept and in the slope of the
regression. Species are ordered from most influential to less influential and
only include species with a standardised difference > \code{cutoff}.

\code{influ.model.estimates}: A data frame with all simulation
estimates. Each row represents a deleted species. Reported are the calculated
regression intercept (\code{intercept}), difference between simulation
intercept and full model intercept (\code{DFintercept}), the standardised
difference (\code{sDFintercept}), the percentage change in intercept compared
to the full model (\code{intercept.perc}) and intercept p-value
(\code{pval.intercept}). All of these are also reported for the regression
slope (\code{DFslope} etc.). Additonally, model aic value (\code{AIC}) and
the optimised value (\code{optpar}) of the phylogenetic parameter
(e.g. \code{kappa} or \code{lambda}, depends on phylogeneticmodel used) are
reported.

\code{data}: Original full dataset.

\code{errors}: Species where deletion resulted in errors.
}
\description{
Performs leave-one-out deletion analyis for phylogenetic linear regression,
and detects influential species.
}
\details{
This function sequentially removes one species at a time, fits a phylogenetic
linear regression model using \code{\link[phylolm]{phylolm}}, stores the
results and detects influential species.

All phylogenetic models from \code{phylolm} can be used, i.e. \code{BM},
\code{OUfixedRoot}, \code{OUrandomRoot}, \code{lambda}, \code{kappa},
\code{delta}, \code{EB} and \code{trend}. See ?\code{phylolm} for details.

\code{influ_phylolm} detects influential species based on the standardised
difference in intercept and/or slope when removing a given species compared
to the full model including all species. Species with a standardised difference
above the value of \code{cutoff} are identified as influential. The default
value for the cutoff is 2 standardised differences change.

Currently, this function can only implement simple linear models (i.e. \eqn{trait~
predictor}). In the future we will implement more complex models.

Output can be visualised using \code{sensi_plot}.
}
\examples{
\dontrun{
library(sensiPhy);library(phylolm)

#Generate a random tree
set.seed(2468)
tree <- rtree(100)

#Generate random predictor variable (pred), evolving according to a BM model.
pred<- rTraitCont(tree,root.value=0,sigma=1,model="BM")

#Generate two continous traits, one evolving highly correlated with the
#predictor (trait 1), and one evolving more randomly (trait 2)
cont_trait1 <- pred + rTraitCont(tree,model="BM",sigma=0.1)
cont_trait2 <- pred + rTraitCont(tree,model="BM",sigma=10)

#Generate two binary traits, one highly correlated to pred (trait 1), the other less.
bin_trait1<-rbinTrait(n=1,tree,beta=c(-1,0.5),alpha=0.1,
                     X=cbind(rep(1,length(tree$tip.label)),pred))
bin_trait2<-rbinTrait(n=1,tree,beta=c(-1,0.5),alpha=5,
                      X=cbind(rep(1,length(tree$tip.label)),pred))
dat<-data.frame(pred,cont_trait1,cont_trait2,bin_trait1,bin_trait2)

#Determine influential species for both regressions.
fit1<-influ_phylolm(cont_trait1~pred,data = dat,phy = tree)
fit2<-influ_phylolm(cont_trait2~pred,data = dat,phy = tree)

#For purposes of comparison the full model output from phylolm:
summary(phylolm(cont_trait1~pred,data = dat,phy = tree,model = "lambda"))
summary(phylolm(cont_trait2~pred,data = dat,phy = tree,model = "lambda"))
}
}
\author{
Gustavo Paterno & Gijsbert D.A. Werner
}
\references{
Here still: reference to phylolm paper + our own?
}
\seealso{
\code{\link[phylolm]{phylolm}}, \code{\link{samp_phylolm}},
\code{\link{influ_phyloglm}},\code{\link{sensi_plot}}
}

