% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intra_phyloglm.R
\name{intra_phyglm}
\alias{intra_phyglm}
\title{Intraspecific variability - Phylogenetic Logistic Regression}
\usage{
intra_phyglm(formula, data, phy, Vx = NULL, Vy = NULL, times = 2,
  distrib = "uniform", btol = 50, track = TRUE, ...)
}
\arguments{
\item{formula}{The model formula: \code{response~predictor}.}

\item{data}{Data frame containing species traits with species as row names.}

\item{phy}{A phylogeny (class 'phylo', see ?\code{ape}).}

\item{Vx}{Name of the column containing the standard deviation or the standard error of the predictor
variable. When information is not available for one taxon, the value can be 0 or \code{NA}}

\item{Vy}{Name of the column containing the standard deviation or the standard error of the response
variable. When information is not available for one taxon, the value can be 0 or \code{NA}}

\item{times}{Number of times to repeat the analysis generating a random value for the predictor variable.
If NULL, \code{times} = 2}

\item{distrib}{A character string indicating which distribution to use to generate a random value for the response
and/or predictor variables. Default is normal distribution: "normal" (function \code{\link{rnorm}}).
Uniform distribution: "uniform" (\code{\link{runif}})
Warning: we recommend to use normal distribution with Vx or Vy = standard deviation of the mean.}

\item{btol}{Bound on searching space. For details see \code{phyloglm}}

\item{track}{Print a report tracking function progress (default = TRUE)}

\item{...}{Further arguments to be passed to \code{phyloglm}}
}
\value{
The function \code{intra_phyglm} returns a list with the following
components:

\code{formula}: The formula

\code{data}: Original full dataset

\code{model_results}: Coefficients, aic and the optimised
value of the phylogenetic parameter (e.g. \code{lambda}) for each regression.

\code{N.obs}: Size of the dataset after matching it with tree tips and removing NA's.

\code{stats}: Statistics for model parameters. \code{sd_intra} is the standard deviation
due to intraspecific variation. \code{CI_low} and \code{CI_high} are the lower and upper limits
of the 95% confidence interval.
}
\description{
Performs Phylogenetic logistic regression evaluating
intraspecific variability.
}
\details{
This function fits a phylogenetic logistic regression model using \code{\link[phylolm]{phyloglm}}.
The regression is repeated \code{times} times. At each iteration the functions generates for each row in the dataset
a random value in the normal or uniform distribution.
Warning: if predictor and/or response variables are log-transformed, please make sure that
Vx and/or Vy are also in a log-scale.

All phylogenetic models from \code{phyloglm} can be used, i.e. \code{BM},
\code{OUfixedRoot}, \code{OUrandomRoot}, \code{lambda}, \code{kappa},
\code{delta}, \code{EB} and \code{trend}. See ?\code{phyloglm} for details.

Currently, this function can only implement simple logistic models (i.e. \eqn{trait~
predictor}). In the future we will implement more complex models.

Output can be visualised using \code{sensi_plot}.
}
\author{
Caterina Penone & Pablo Ariel Martinez
}
\references{
Here still: reference to phylolm paper + our own?
}
\seealso{
\code{\link{sensi_plot}}
}

