% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_samp_phylolm.R
\name{plot_samp_phylolm}
\alias{plot_samp_phylolm}
\title{Graphical diagnostics for \code{samp_phylolm}}
\usage{
plot_samp_phylolm(x, graphs = "all", param = "slope")
}
\arguments{
\item{x}{output from \code{samp_phylolm}}

\item{graphs}{choose which graph should be printed on the output ("all", 1,2,3 or 4)}

\item{param}{choose which parameter ("intercept" or "slope") should be printed}
}
\description{
\code{plot_samp_phylolm} Plot results from \code{samp_phylolm},
}

