% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_tree.intra_phylolm.R
\name{plot_tree.intra_phylolm}
\alias{plot_tree.intra_phylolm}
\title{Graphical diagnostics for \code{tree_phylolm} and \code{intra_phylolm}}
\usage{
plot_tree.intra_phylolm(x, graphs = "all", param = NULL)
}
\arguments{
\item{x}{output from \code{tree_phylolm} or \code{intra_phylolm}}

\item{graphs}{choose which graph should be printed on the output ("all", 1, 2 or 3)}

\item{param}{choose which parameter ("intercept" or "slope" should be printed)}
}
\description{
\code{plot_tree.intra_phylolm} Plot results from \code{tree_phylolm} and
\code{intra_phylolm}
}
\details{
For 'x' from tree_phylolm or intra_phylolm:

Graphs 1 and 2: Distribution of estimated slopes and intercepts for each tree (for \code{tree_phylolm}) or
value generated within a given interval (\code{intra_phylolm})
Red vertical line represents the mean slope or intercept for all models.

Graph 3: Scatterplot with mean regression (black line) and standard deviation of the regression (blue dotted lines).
}
\examples{
\dontrun{
library(sensiPhy)

}
}
\author{
Caterina Penone and Gustavo Paterno
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[sensiPhy]{tree_phylolm}}
\code{\link[sensiPhy]{intra_phylolm}}
}

