% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predtreeVar_pgls.R
\name{predtreeVar_pgls}
\alias{predtreeVar_pgls}
\title{PGLS analysis accounting for variability in predictors and trees.}
\usage{
predtreeVar_pgls(resp, pred, se.pred = NA, tree, ntree = 1, npred = 1,
  method = c("normal", "uniform"), taxa.col, lambda = "ML")
}
\arguments{
\item{resp}{Numeric vector containing the response variable}

\item{pred}{Vector containing the predictor variable}

\item{se.pred}{Vector containing the standard error of \code{pred}}

\item{tree}{A tree or list of tree of class \code{Phylo} or \code{multiPhylo}}

\item{ntree}{If TRUE or class(tree)=multiPhylo, the number of times to repeat the analysis with n different
trees picked randomly in the multiPhylo file. If NULL, \code{ntree}=1}

\item{npred}{If TRUE, the number of times to repeat the analysis generating a random value in the interval
[\code{pred}-\code{se.pred},\code{pred}+\code{se.pred}]. If NULL, \code{npred}=1}

\item{method}{A character string indicating which method to use to generate a random value in the interval
[\code{pred} - \code{se.pred}, \code{pred} + \code{se.pred}]. Default is normal distribution: "normal"
(function \code{\link{rnorm}}), uniform distribution is "uniform" (\code{\link{runif}})}

\item{taxa.col}{A character vector of taxa names that will be used to match data rows to phylogeny tips.}

\item{lambda}{A value for the lambda transformation. If NULL, \code{lambda}="ML"}
}
\value{
The function \code{predtreeVar_pgls} returns a list with the following
components:

\code{model_results} Model parameters for each iteration

\code{stats} Mean and sd of each parameter of the model
}
\description{
\code{predtreeVar_pgls} performs \code{pgls} analysis accounting for variability in predictors and trees.
It picks n times a random value within an interval for the predictor variable and
it repeats the analysis m times for m trees in a \code{multiPhylo} file.
The output gives a mean and sd of the model parameters.
}
\details{
This functions only works for simple linear regression \eqn{y = bx +
a}. Future implementation will deal with more complex models.
}
\section{Warning}{
 This code is not fully checked. Please be aware.
If \code{ntree} and \code{npred} are set to 1, the function computes a classic pgls.
}
\examples{
to do :)
}
\seealso{
\code{\link{pgls}}, \code{\link{comparative.data}}, \code{\link{runif}}, \code{\link{rnorm}}
}

