% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/samp_pgls.R
\name{samp_pgls}
\alias{samp_pgls}
\title{Sampling effort analysis for pgls regression.}
\usage{
samp_pgls(formula, data, times = 20, breaks = seq(0.1, 0.7, 0.1),
  lambda = "ML")
}
\arguments{
\item{formula}{A model formula}

\item{data}{A 'comparative.data' object containing the covariance matrix and
data to be used in the model.}

\item{times}{The number of times to repeat each simulation (per
\code{breaks}) interval.}

\item{breaks}{Percentage intervals to remove species. For example:
\code{breaks = c(.1,.2,.3)},removes 10,20 and 30 percentage of species at
random in each simulation.}

\item{lambda}{A value for the lambda transformation. If NULL, lambda= "ML}
}
\value{
The function \code{samp_pgls} returns a list with the following
  components:

\code{model_estimates} Full model estimates

\code{beta95_IC} Full model beta 95 confidence interval

\code{results} A data frame with all simulation estimates.

\code{power_analysis} A data frame with power analysis for each
}
\description{
\code{samp_pgls} performs sampling effort diagnostics for \code{pgls}
regressions. It removes species at random, fits a pgls model without the
species and store the results of the model estimates. The percentage of
species removed is specified with \code{breaks} and the number of simulations
per break is defined by \code{times}.
}
\details{
This functions only works for simple linear regression \eqn{y = bx +
  a}. Future implementation will deal with more complex models.
}
\section{Warning}{
 This code is note fully checked. Please be aware.
}
\seealso{
\code{\link{pgls}}, \code{\link{influ_pgls}}
}

