% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samp_phylolm.R
\name{samp_phylm}
\alias{samp_phylm}
\title{Sensitivity Analysis Species Sampling  - Phylogenetic Linear Regression}
\usage{
samp_phylm(formula, data, phy, times = 20, breaks = seq(0.1, 0.7, 0.1),
  model = "lambda", track = TRUE, ...)
}
\arguments{
\item{formula}{The model formula}

\item{data}{Data frame containing species traits with row names matching tips
in \code{phy}.}

\item{phy}{A phylogeny (class 'phylo') matching \code{data}.}

\item{times}{The number of times species are randomly deleted for each
\code{break}.}

\item{breaks}{A vector containing the percentages of species to remove.}

\item{model}{The phylogenetic model to use (see Details). Default is \code{lambda}.}

\item{track}{Print a report tracking function progress (default = TRUE)}

\item{...}{Further arguments to be passed to \code{phylolm}}
}
\value{
The function \code{samp_phylm} returns a list with the following
components:

\code{formula}: The formula

\code{full.model.estimates}: Coefficients, aic and the optimised
value of the phylogenetic parameter (e.g. \code{lambda} or \code{kappa}) for
the full model without deleted species.

\code{samp.model.estimates}: A data frame with all simulation
estimates. Each row represents a model rerun with a given number of species
\code{n.remov} removed, representing \code{n.percent} of the full dataset.
Reported are the calculated regression intercept (\code{intercept}),
difference between simulation intercept and full model intercept (\code{DFintercept}),
the percentage change in intercept compared to the full model (\code{intercept.perc})
and intercept p-value (\code{pval.intercept}). All of these are also reported
for the regression slope (\code{DFslope} etc.). Additonally, model aic value
(\code{AIC}) and the optimised value (\code{optpar}) of the phylogenetic
parameter (e.g. \code{kappa} or \code{lambda}, depends on phylogeneticmodel
used) are reported.

\code{sign.analysis} For each break (i.e. each percentage of species
removed) this reports the percentage of statistically signficant (at p<0.05)
intercepts (\code{perc.sign.intercept}) over all repititions as well as the
percentage of statisticaly significant (at p<0.05) slopes (\code{perc.sign.slope}).

\code{data}: Original full dataset.
}
\description{
Performs analyses of sensitivity to species sampling by randomly removing
species and detecting the effects on parameter estimates in a phylogenetic
linear regression.
}
\details{
This function randomly removes a given percentage of species (controlled by
\code{breaks}) from the full phylogenetic linear regression, fits a phylogenetic
linear regression model without these species using \code{\link[phylolm]{phylolm}},
repeats this many times (controlled by \code{times}), stores the results and
calculates the effects on model parameters.

All phylogenetic models from \code{phylolm} can be used, i.e. \code{BM},
\code{OUfixedRoot}, \code{OUrandomRoot}, \code{lambda}, \code{kappa},
\code{delta}, \code{EB} and \code{trend}. See ?\code{phylolm} for details.

Currently, this function can only implement simple linear models (i.e. \eqn{trait~
predictor}). In the future we will implement more complex models.

Output can be visualised using \code{sensi_plot}.
}
\author{
Gustavo Paterno & Gijsbert D.A. Werner
}
\references{
Ho, L. S. T. and Ane, C. 2014. "A linear-time algorithm for
Gaussian and non-Gaussian trait evolution models". Systematic Biology 63(3):397-408.
}
\seealso{
\code{\link[phylolm]{phylolm}}, \code{\link{samp_phyglm}},
\code{\link{influ_phylm}},\code{\link{sensi_plot}}
}

