% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samp_phyloglm.R
\name{samp_phyloglm}
\alias{samp_phyloglm}
\title{Sampling effort analysis for phylogenetic generalized linear models (logistic regression).}
\usage{
samp_phyloglm(formula, data, phy, times = 20, breaks = seq(0.1, 0.7, 0.1),
  btol = 50)
}
\arguments{
\item{formula}{A model formula}

\item{data}{A data frame containing variables that can be attributed to the taxa
at the tips of a phylogeny}

\item{phy}{A phylogeny (class 'phylo') to be matched to the data above}

\item{times}{The number of times to repeat each simulation (per
\code{breaks}) interval.}

\item{breaks}{Percentage intervals to remove species. For example:
\code{breaks = c(.1,.2,.3)},removes 10, 20 and 30 percent of species at
random in each simulation.}

\item{btol}{Bound on the linear predictor to bound the searching space (see ?phyloglm)}
}
\value{
The function \code{samp_phyloglm} returns a list with the following
  components:

\code{model_estimates} Full model estimates. This contains beta and intercept for the full model and associated p-values, as well as the phylogenetic correlation parameter alpha from \code{phyloglm}.

\code{results} A data frame with all simulation estimates.

\code{power_analysis} A data frame with power analysis for each

\code{data} Original dataset
}
\description{
\code{samp_phyloglm} performs sample size sensitivity analysis for logistic regresssions using \code{phyloglm}.
It removes species at random, fits a phyloglm model without the
species and stores the results of the model estimates. The percentage of
species removed is specified with \code{breaks} and the number of simulations
per break is defined by \code{times}.
}
\details{
This functions currently only works for logistic regressions, Poisson regressions
will be implemented later (see R-package phylolm, Ives and Garland 2002 and Ives and Garland 2010).
Also it currently only uses logistic_MPLE as a method, and can only deal with simple logistic regression models.
}
\section{Warning}{
 This code is not fully checked. Please be aware.
}
\examples{
library(caper);library(ggplot2);library(gridExtra);library(phylolm)
data(shorebird)
sp.ord <- match(shorebird.tree$tip.label, rownames(shorebird.data))
shorebird.data <- shorebird.data[sp.ord,]
#Create a binary variable (large egg / small egg), for illustration purposes.
mean(shorebird.data$Egg.Mass)
shorebird.data$Egg.Mass.binary<-ifelse(shorebird.data$Egg.Mass>30,1,0) #Turn egg mass into a binary variable
table(shorebird.data$Egg.Mass.binary) #Mostly small eggs.
#Run samp_phyloglm
samp_phyloglm1<-samp_phyloglm(formula = Egg.Mass.binary~M.Mass,data = shorebird.data,phy=shorebird.tree,btol = 50)
sensi_plot(samp_phyloglm1)
}
\seealso{
\code{\link{pgls}}, \code{\link{influ_pgls}}, \code{\link{samp_pgls}}, \code{\link{samp_phyloglm}}
}

