% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samp_regression.R
\name{samp_regression}
\alias{samp_regression}
\title{Sampling effort analysis for gls phylogenetic regression.}
\usage{
samp_regression(formula, data, phy, times = 50, breaks = seq(0.1, 0.5, 0.1))
}
\arguments{
\item{formula}{A model formula}

\item{data}{A data frame containing variables that can be attributed to the taxa
at the tips of a phylogeny}

\item{phy}{A phylogeny (class 'phylo') to be matched to the data above}

\item{times}{The number of times to repeat each simulation (per
\code{breaks}) interval.}

\item{breaks}{Percentage intervals to remove species. For example:
\code{breaks = c(.1,.2,.3)},removes 10,20 and 30 percentage of species at
random in each simulation.}
}
\value{
The function \code{samp_gls} returns a list with the following
  components:

\code{model_estimates} Full model estimates

\code{intervals} Full model 95% confidence interval

\code{results} A data frame with all simulation estimates.

\code{power_analysis} A data frame with power analysis

\code{data} Original dataset
}
\description{
\code{samp_regression} performs sample size sensitive analysis for \code{pgls}
regressions with continuous variables. It removes species at random, fits a pgls model without the
species and store the results of the model estimates. The percentage of
species removed is specified with \code{breaks} and the number of simulations
per break is defined by \code{times}.
}
\details{
This functions only works with continuous variables on Y and X. More
then one explanatory variable and interactions are allowed in the model.
Future implementation will deal categorical explanatory variables.
}
\section{Warning}{
 This code is note fully checked. Please be aware.
}
\seealso{
\code{\link{pgls}}, \code{\link{influ_pgls}}, \code{\link{samp_pgls}}
}

