% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi_plot.clade.physig.R
\name{sensi_plot.clade.physig}
\alias{sensi_plot.clade.physig}
\title{Graphical diagnostics for class 'clade.physig'}
\usage{
\method{sensi_plot}{clade.physig}(x, clade = NULL, ...)
}
\arguments{
\item{x}{output from \code{influ_physig}}

\item{clade}{The name of the clade to be evaluated (see details)}

\item{...}{further arguments to methods.}
}
\description{
\code{plot_clade_physig} Plot results from \code{clade_physig}
}
\details{
For 'x' from clade_physig:

\strong{Graph 1:} Distribution of the simulated signal estimates (Null distribution
for a given clade sample size).
The red dashed line represents the estimated signal for the reduced data 
(without the focal clade) and the black line represents the signal estimate
 for the full data.
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[phytools]{phylosig}}
}
\author{
Gustavo Paterno
}
