% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi_plot.influ.physig.R
\name{sensi_plot.influ.physig}
\alias{sensi_plot.influ.physig}
\title{Graphical diagnostics for class 'influ.physig'}
\usage{
\method{sensi_plot}{influ.physig}(x, graphs = "all", ...)
}
\arguments{
\item{x}{output from \code{influ_physig}}

\item{graphs}{choose which graph should be printed on the output ("all", 1, 2)}
}
\description{
\code{sensi_plot.influ_physig} Plot results from \code{influ_physig}
}
\details{
For 'x' from influ_physig:

\strong{Graph 1:} Distribution of estimated phylogenetic signal (K or lambda) for each 
simulation (leave-one-out deletion). Dashed red vertical line represents the original
 estimate of phylogenetic signal with the full data (with all species). 

strong{Graph 2:} Distribution of P values for the phylogenetic signal (K or lambda) 
for each simulation (leave-one-out deletion). Red vertical line represents 
the alpha significance level = 0.05.
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[phytools]{phylosig}}
}
\author{
Gustavo Paterno
}
