% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi_plot.tree.physig.R
\name{sensi_plot.intra.physig}
\alias{sensi_plot.intra.physig}
\title{Graphical diagnostics for class 'intra.physig'}
\usage{
\method{sensi_plot}{intra.physig}(x, graphs = "all", ...)
}
\arguments{
\item{x}{output from \code{intra_physig}}

\item{graphs}{choose which graph should be printed in the output ("all", 1 or 2)}

\item{...}{further arguments to methods}
}
\description{
\code{sensi_plot_intra.physig} Plot results from \code{intra_physig}.
}
\details{
For 'x' from \code{intra_physig}

Graphs 1: Distribution of estimated phylogenetic signal (lambda or K) for each simulation
Red vertical line represents the average signal among all estimates. 

Graph 2: Distribution of p-values for the phylogenetic signal (K or lambda) 
for each simulation. Red vertical line represents the alpha significance level = 0.05.
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[sensiPhy]{intra_phylm}}
\code{\link[sensiPhy]{intra_phylm}}
}
\author{
Caterina Penone and Gustavo Paterno
}
