% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi_plot.samp.physig.R
\name{sensi_plot.samp.physig}
\alias{sensi_plot.samp.physig}
\title{Graphical diagnostics for class 'samp.physig'}
\usage{
\method{sensi_plot}{samp.physig}(x, graphs = "all", ...)
}
\arguments{
\item{x}{output from \code{samp_physig}}

\item{graphs}{choose which graph should be printed on the output ("all", 1,2 and 3 )}

\item{...}{further arguments to methods}
}
\description{
\code{plot_samp_phylm} Plot results from \code{samp_physig}
}
\details{
For 'x' from samp_physig:

\strong{Graph 1:} Estimated phylogenetic signal for each simulation across  
percentages of species removed. Colours represent percentage 
of change in comparison with the full data estimate (blue = lower than 5, orange = 
between 5 and 10 and red = higher than 10).
The red horizontal line represents the original phylogenetic signal 
estimated from the full model (with all species). 

\strong{Graph 2:} The proportion of estimated signal in each category 
across the percentage of species removed.

\strong{Graph 3:} The percentage of significant signal estimates across the 
percentage of species removed.
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[sensiPhy]{samp_phylm}}
\code{\link[sensiPhy]{samp_physig}}
}
\author{
Gustavo Paterno
}
