% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi_plot.sensiClade.R
\name{sensi_plot.sensiClade.TraitEvol}
\alias{sensi_plot.sensiClade.TraitEvol}
\title{Graphical diagnostics for class 'sensiClade.TraitEvol'}
\usage{
\method{sensi_plot}{sensiClade.TraitEvol}(x, clade = NULL, graph = "all", ...)
}
\arguments{
\item{x}{output from \code{clade_discrete} or \code{clade_continuous}}

\item{clade}{The name of the clade to be evaluated (see details)}

\item{graph}{The graph to be printed. Default \code{all}, or 
for \code{clade_continuous} set \code{sigsq} or \code{optpar} 
and for \code{clade_discrete} set\code{q12} or \code{q21}.}

\item{...}{further arguments to methods.}
}
\description{
Plot results from \code{clade_discrete} and \code{clade_continuous}
}
\details{
For 'x' from clade_discrete or clade_continuous:

\strong{Graph 1:} Distribution of the simulated parameter estimates (Null distribution
for a given clade sample size).
The red dashed line represents the estimated signal for the reduced data 
(without the focal clade) and the black line represents the signal estimate
 for the full data.
}
\seealso{
\code{\link[sensiPhy]{clade_continuous}} or \code{\link[sensiPhy]{clade_discrete}}
}
\author{
Gustavo Paterno & Gijsbert Werner
}
