% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_tree.intra_phylolm.R
\name{sensi_plot.sensiIntra}
\alias{sensi_plot.sensiIntra}
\title{Graphical diagnostics for class 'sensiIntra'}
\usage{
\method{sensi_plot}{sensiIntra}(x, graphs = "all")
}
\arguments{
\item{x}{output from \code{tree_phylolm}, \code{tree_phyloglm},
\code{intra_phylolm} or \code{intra_phyloglm}}

\item{graphs}{choose which graph should be printed in the output ("all", 1, 2 or 3)}

\item{...}{further arguments to methods}
}
\description{
\code{plot_tree.intra_phylolm} Plot results from \code{tree_phylolm},
\code{intra_phylolm} and \code{intra_phyloglm}
}
\details{
For 'x' from \code{tree_phylolm}, \code{tree_phyloglm},
\code{intra_phylolm} or \code{intra_phyloglm}:

Graphs 1 and 2: Distribution of estimated slopes and intercepts for each tree (for \code{tree_phylolm}) or
value generated within a given interval (\code{intra_phylolm})
Red vertical line represents the mean slope or intercept for all models.

Graph 3: Scatterplot with mean regression (black line) and standard deviation of the regression (blue dotted lines).
}
\examples{
\dontrun{
library(sensiPhy)

}
}
\author{
Caterina Penone and Gustavo Paterno
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[sensiPhy]{tree_phylolm}}
\code{\link[sensiPhy]{intra_phylolm}}
}

