% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi_plot.sensiINTER_Influ.R
\name{sensi_plot.sensiIntra_Influ}
\alias{sensi_plot.sensiIntra_Influ}
\title{Graphical diagnostics for class 'sensiIntra_Influ'}
\usage{
\method{sensi_plot}{sensiIntra_Influ}(x, graphs = "all", ...)
}
\arguments{
\item{x}{output from \code{influ_phylm} or \code{intra_influ_phyglm}}

\item{graphs}{choose which graph should be printed on the output ("all", 1,2,3 or 4)}

\item{...}{further arguments to methods}
}
\description{
\code{sensi_plot.intra_influ} Plot results from \code{intra_influ_phylm} and 
\code{intra_influ_phyglm}
}
\details{
For 'x' from intra_influ_phylm or intra_influ_phyglm:

\strong{Graph 1:} Most common influential species on model estimates.
 Percentage of iterations (n.tree or n.intra) where the removal of individual
  species caused significant change in model estimate (sDIFestimate > cutoff). 

\strong{Graph 2:} Shift on model estimate after species removal for most influential
 species. Small red dots represent individual reruns between different trees or 
 simulations while large red dots represent the average DIFestimate among all 
 iterations.
}
\seealso{
\code{\link[sensiPhy]{intra_influ_phylm}}; 
\code{\link[sensiPhy]{intra_influ_phyglm}}
}
\author{
Gustavo Paterno, Caterina Penone
}
