% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi_plot.sensiINTER_Samp.R
\name{sensi_plot.sensiIntra_Samp}
\alias{sensi_plot.sensiIntra_Samp}
\title{Graphical diagnostics for class 'sensiIntra_Samp' and 'sensiTree_samp'}
\usage{
\method{sensi_plot}{sensiIntra_Samp}(x, graphs = "all", ...)
}
\arguments{
\item{x}{output from \code{sensiIntra_Samp_phylm} or \code{sensiIntra_Samp_phyglm}}

\item{graphs}{choose which graph should be printed on the output ("all", 1,2,3 or 4)}

\item{...}{further arguments to methods}

\item{param}{choose which model parameter should be ploted  ("intercept" or "estimate")}
}
\description{
\code{sensi_plot.sensiIntra_Samp} Plot results from \code{sensiIntra_Samp_phylm} and
\code{sensiIntra_Samp_phyglm}
}
\details{
For 'x' from sensiIntra_Samp_phylm or sensiIntra_Samp_phyglm:

\strong{Graph 1:} : Estimated estimates for each simulation across percentages of 
species removed. Small red dots represent individual estimates for each iteration 
(tree or intra) across percentages of removed species. Large dots represent the
 average among these estimates within each iteration and percentage of species removal.
  Different estimates within the same percentage interval represent different 
  iterations (tree or intra). 

\strong{Graph 2:} : The percentage of significant estimates (p < 0.05) across the 
percentage of species removed. Small red dots represent individual 
iterations (tree or intra) while large dots represent the average among these iterations.
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[sensiPhy]{samp_phylm}}
\code{\link[sensiPhy]{samp_phyglm}}
}
\author{
Gustavo Paterno
}
