% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi_plot.sensiSamp.R
\name{sensi_plot.sensiSamp.TraitEvol}
\alias{sensi_plot.sensiSamp.TraitEvol}
\title{Graphical diagnostics for class 'sensiSamp.TraitEvol'}
\usage{
\method{sensi_plot}{sensiSamp.TraitEvol}(x, graphs = "all", ...)
}
\arguments{
\item{x}{output from \code{samp_continuous} or \code{samp_continuous}}

\item{graphs}{choose which graph should be printed on the output ("all", 1,2,3 or 4)}

\item{...}{further arguments to methods}
}
\description{
\code{sensi_plot.sensiSamp.TraitEvol} Plot results from \code{samp_continuous} and
\code{samp_discrete}
}
\details{
For 'x' from samp_continuous or samp_discrete:

\strong{Graph 1:} Estimated q12 (discrete) or sigsq (discrete) for each simulation across  
percentages of species removed. Colours represent percentage 
of change in comparison with the full model (blue = lower than 5, orange = 
between 5 and 10 and red = higher than 10).The red horizontal line represents 
the original value from the full model (with all species). 

\strong{Graph 2:} The proportion of estimated  q12 (discrete) or sigsq (discrete) in each category 
across the percentage of species removed.

\strong{Graph 3:} Estimated q21 for each simulation across the percentage of species removed  
(only for \code{samp_discrete}).

\strong{Graph 4:} The percentage of significant q21 across the 
percentage of species removed (only for \code{samp_discrete}).
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[sensiPhy]{samp_continuous}}
\code{\link[sensiPhy]{samp_discrete}}
}
\author{
Gustavo Paterno & Gijsbert Werner
}
