% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_tree.intra_phylolm.R
\name{sensi_plot.sensiTree}
\alias{sensi_plot.sensiTree}
\title{Graphical diagnostics for class 'sensiTree'}
\usage{
\method{sensi_plot}{sensiTree}(x, graphs = "all", ...)
}
\arguments{
\item{x}{output from \code{tree_phylolm}, \code{tree_phyloglm},
\code{intra_phylm} or \code{intra_phyglm}}

\item{graphs}{choose which graph should be printed in the output ("all", 1, 2 or 3)}

\item{...}{further arguments to methods}
}
\description{
\code{plot_tree.intra_phylm} Plot results from \code{tree_phylolm},
\code{intra_phylm} and \code{intra_phyglm}
}
\details{
For 'x' from \code{tree_phylolm}, \code{tree_phyloglm},
\code{intra_phylm} or \code{intra_phyglm}:

Graphs 1 and 2: Distribution of estimated slopes and intercepts for each tree (for \code{tree_phylolm}) or
value generated within a given interval (\code{intra_phylm})
Red vertical line represents the mean slope or intercept for all models.

Graph 3: Scatterplot with mean regression (black line) and standard deviation of the regression (blue dotted lines).
}
\author{
Caterina Penone and Gustavo Paterno
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[sensiPhy]{tree_phylolm}}
\code{\link[sensiPhy]{intra_phylm}}
}

