% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi_plot.sensiIntra.sensiTree.R
\name{sensi_plot.sensiTree_Intra}
\alias{sensi_plot.sensiTree_Intra}
\title{Graphical diagnostics for class 'sensiIntra_Tree'}
\usage{
\method{sensi_plot}{sensiTree_Intra}(x, graphs = "all", uncer.type = "all",
  ...)
}
\arguments{
\item{x}{output from \code{tree_intra_phylm}, \code{tree_intra_phyglm}}

\item{graphs}{choose which graph should be printed in the output ("all", 1, 2 or 3)}

\item{uncer.type}{chosse which uncertainty type should be printed ("all", "intra", "tree")}

\item{...}{further arguments to methods}
}
\description{
\code{plot_tree.intra_phylm} Plot results from \code{tree_intra_phylm},
\code{tree_intra_phylm} and \code{tree_intra_phyglm}
}
\details{
For 'x' from \code{tree_intra_phylm} or \code{tree_intra_phyglm}:

\strong{Graphs 1 and 2:} Distribution of estimated estimates and intercepts for each tree (for \code{tree_phylm}) or 
value generated within a given interval (\code{interaction_intra_tree_phylm})
Red vertical line represents the mean estimate or intercept for all models. 

\strong{Graph 3:} Scatterplot with mean regression (black line) and standard deviation of the regression (blue dotted lines).
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[sensiPhy]{tree_phylm}}
\code{\link[sensiPhy]{intra_phylm}}
}
\author{
Caterina Penone and Gustavo Paterno
}
