% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi_plot.sensiINTER_Samp.R
\name{sensi_plot.sensiTree_Samp}
\alias{sensi_plot.sensiTree_Samp}
\title{Graphical diagnostics for class 'sensiTree_samp'}
\usage{
\method{sensi_plot}{sensiTree_Samp}(x, graphs = "all", ...)
}
\arguments{
\item{x}{output from \code{sensiTree_Samp_phylm} or \code{sensiTree_Samp_phyglm}}

\item{graphs}{choose which graph should be printed on the output ("all", 1,2,3 or 4)}

\item{...}{further arguments to methods}
}
\description{
\code{sensi_plot.sensiTree_Samp} Plot results from \code{sensiTree_Samp_phylm} and
\code{sensiTree_Samp_phyglm}
}
\details{
For 'x' from sensiTree_Samp_phylm or sensiTree_Samp_phyglm:

\strong{Graph 1:} : Estimated estimates for each simulation across percentages of 
species removed. Small red dots represent individual estimates for each iteration 
(tree or intra) across percentages of removed species. Large dots represent the
 average among these estimates within each iteration and percentage of species removal.
  Different estimates within the same percentage interval represent different 
  iterations (tree or intra). 

\strong{Graph 2:} : The percentage of significant estimates (p < 0.05) across the 
percentage of species removed. Small red dots represent individual 
iterations (tree or intra) while large dots represent the average among these iterations.
}
\note{
If model = "BM", only plots 1, 2 and 4 are printed. Plot 3, phylogenetic
 model parameter is not available for model = "BM"
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[sensiPhy]{samp_phylm}}
\code{\link[sensiPhy]{samp_phyglm}}
}
\author{
Gustavo Paterno
}
