% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi_plot.tree.bd.R
\name{sensi_plot.tree.bd}
\alias{sensi_plot.tree.bd}
\title{Graphical diagnostics for class 'tree.physig'}
\usage{
\method{sensi_plot}{tree.bd}(x, graphs = "all", ...)
}
\arguments{
\item{x}{output from \code{tree_bd}}

\item{graphs}{choose which graph should be printed in the output ("all", 1 or 2)}

\item{...}{further arguments to methods}
}
\description{
\code{sensi_plot_tree.bd} Plot results from \code{tree_bd}.
}
\details{
For 'x' from \code{tree_bd}

\strong{Graphs 1:} Distribution of estimated diversification or speciation rate
 for each tree. Red vertical line represents the average signal among all estimates. 
 
\strong{Graphs 2:} Estimates across each phylogenetic tree.
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[sensiPhy]{tree_phylm}}
\code{\link[sensiPhy]{intra_phylm}}
}
\author{
Caterina Penone and Gustavo Paterno
}
