% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi_plot.tree.physig.R
\name{sensi_plot.tree.physig}
\alias{sensi_plot.tree.physig}
\title{Graphical diagnostics for class 'tree.physig'}
\usage{
\method{sensi_plot}{tree.physig}(x, graphs = "all", ...)
}
\arguments{
\item{x}{output from \code{tree_physig}}

\item{graphs}{choose which graph should be printed in the output ("all", 1 or 2)}

\item{...}{further arguments to methods}
}
\description{
\code{sensi_plot_tree.physig} Plot results from \code{tree_physig}.
}
\details{
For 'x' from \code{tree_physig}

\strong{Graphs 1:} Distribution of estimated phylosgentic signal (lambda or K) for each tree
Red vertical line represents the average signal among all estimates. 

\strong{Graph 2:} Distribution of p-values for the phylogenetic signal (K or lambda) 
for each tree. Red vertical line represents the alpha significance level = 0.05.
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[sensiPhy]{tree_phylm}}
\code{\link[sensiPhy]{intra_phylm}}
}
\author{
Caterina Penone and Gustavo Paterno
}
