% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sensi_time.R
\name{sensi_time}
\alias{sensi_time}
\title{Estimates time of samp_pgls and influ_pgls}
\usage{
sensi_time(formula, data, times = 20, breaks = seq(0.1, 0.7, 0.1),
  lambda = "ML")
}
\description{
\code{sensi_time} Estimates simulation time to run \code{samp_pgls} and
\code{influ_pgls}.
}

