% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sensiPhy.R
\docType{package}
\name{sensiPhy}
\alias{sensiPhy}
\alias{sensiPhy-package}
\title{Sensitivity Analysis for Comparative Methods}
\description{
The functions implemented in \bold{sensiPhy} provide diagnostics for phylogenetic
generalized least squares by various simulations methods: data removal,
leave-one-out deletion, random sampling of phylogenies and data.
A graphical output with several plots for model diagnostic is provided
for each function in the package.
}
\section{Bug reports}{


\url{https://github.com/paternogbc/sensiPhy/issues}.
}

\section{How to contribute?}{


You can find instructions to contribute with \bold{sensiPhy} at
this link: \url{https://github.com/paternogbc/sensiPhy/wiki/How-to-support-sensiPhy}
}
\author{
Gustavo Paterno; Caterina Penone & Gijsbert D.A. Werner
}

