% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree_phylolm.R
\name{tree_phylolm}
\alias{tree_phylolm}
\title{Phylogenetic uncertainty - Phylogenetic Linear Regression}
\usage{
tree_phylolm(formula, data, phy, ntree = 2, model = "lambda",
  track = TRUE, ...)
}
\arguments{
\item{formula}{The model formula}

\item{data}{Data frame containing species traits with species as row names.}

\item{phy}{A phylogeny (class 'multiPhylo', see ?\code{ape}).}

\item{ntree}{Number of times to repeat the analysis with n different trees picked
randomly in the multiPhylo file.
If NULL, \code{ntree} = 2}

\item{model}{The phylogenetic model to use (see Details). Default is \code{lambda}.}

\item{track}{Print a report tracking function progress (default = TRUE)}

\item{...}{Further arguments to be passed to \code{phylolm}}
}
\value{
The function \code{tree_phylolm} returns a list with the following
components:

\code{formula}: The formula

\code{data}: Original full dataset

\code{model_results}: Coefficients, aic and the optimised
value of the phylogenetic parameter (e.g. \code{lambda}) for each regression with a
different phylogenetic tree.

\code{N.obs}: Size of the dataset after matching it with tree tips and removing NA's.

\code{stats}: Statistics for model parameters. code{sd_tree} is the standard deviation
due to phylogenetic uncertainty.

#which example should we put here?
}
\description{
Performs Phylogenetic linear regression evaluating
uncertainty in trees topology.
}
\details{
This function fits a phylogenetic linear regression model using \code{\link[phylolm]{phylolm}}
to n trees, randomly picked in a multiPhylo file.

All phylogenetic models from \code{phylolm} can be used, i.e. \code{BM},
\code{OUfixedRoot}, \code{OUrandomRoot}, \code{lambda}, \code{kappa},
\code{delta}, \code{EB} and \code{trend}. See ?\code{phylolm} for details.

Currently, this function can only implement simple linear models (i.e. \eqn{trait~
predictor}). In the future we will implement more complex models.

Output can be visualised using \code{sensi_plot}.
}
\author{
Caterina Penone & Pablo Ariel Martinez
}
\references{
Here still: reference to phylolm paper + our own?
}
\seealso{
\code{\link{sensi_plot}}
}

