% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clade_phylolm.R
\name{clade_phylm}
\alias{clade_phylm}
\title{Influential clade detection - Phylogenetic Linear Regression}
\usage{
clade_phylm(formula, data, phy, model = "lambda", track = TRUE, clade.col,
  n.species = 5, ...)
}
\arguments{
\item{formula}{The model formula}

\item{data}{Data frame containing species traits with row names matching tips
in \code{phy}.}

\item{phy}{A phylogeny (class 'phylo') matching \code{data}.}

\item{model}{The phylogenetic model to use (see Details). Default is \code{lambda}.}

\item{track}{Print a report tracking function progress (default = TRUE)}

\item{clade.col}{The name of a column in the provided data frame with clades
specification (a character vector with clade names).}

\item{n.species}{Minimum number of species in the clade in order to include
this clade in the leave-one-out deletion analyis. Default is \code{5}.}

\item{...}{Further arguments to be passed to \code{phylolm}}
}
\value{
The function \code{clade_phylm} returns a list with the following
components:

\code{formula}: The formula

\code{full.model.estimates}: Coefficients, aic and the optimised
value of the phylogenetic parameter (e.g. \code{lambda}) for the full model
without deleted species.

\code{clade.model.estimates}: A data frame with all simulation
estimates. Each row represents a deleted clade. Reported are the calculated
regression intercept (\code{intercept}), difference between simulation
intercept and full model intercept (\code{DFintercept}), the percentage change
 in intercept compared to the full model (\code{intercept.perc}) and intercept
 p-value (\code{pval.intercept}). All of these are also reported for the regression
slope (\code{DFslope} etc.). Additonally, model aic value (\code{AIC}) and
the optimised value (\code{optpar}) of the phylogenetic parameter
(e.g. \code{kappa} or \code{lambda}, depends on phylogeneticmodel used) are
reported.

\code{data}: Original full dataset.

\code{errors}: Clades where deletion resulted in errors.
}
\description{
Estimate the impact on model estimates phylogenetic linear regression after
removing clades from the analysis.
}
\details{
This function sequentially removes one clade at a time, fits a phylogenetic
linear regression model using \code{\link[phylolm]{phylolm}} and stores the
results. The impact of of a specific clade on model estimates is calculated by the
 comparison between the full model (with all species) and the model without
 the species belonging to a clade.

All phylogenetic models from \code{phylolm} can be used, i.e. \code{BM},
\code{OUfixedRoot}, \code{OUrandomRoot}, \code{lambda}, \code{kappa},
\code{delta}, \code{EB} and \code{trend}. See ?\code{phylolm} for details.

\code{clade_phylm} detects influential clades based on
difference in intercept and/or slope when removing a given clade compared
to the full model including all species.

Currently, this function can only implement simple linear models (i.e.
\eqn{y = a + bx}). In the future we will implement more complex models.

Output can be visualised using \code{sensi_plot}.
}
\examples{
\dontrun{
library(sensiPhy);library(phylolm)

#Generate a random tree
set.seed(2468)
tree <- rtree(100)

#Generate random predictor variable (pred), evolving according to a BM model.
pred<- rTraitCont(tree,root.value=0,sigma=1,model="BM")

#Generate one continous traits
cont_trait <- pred + rTraitCont(tree,model="BM",sigma=2.5)
fam <- rep(c("fam1","fam2","fam3","fam4","fam5"),each=20)
dat<-data.frame(pred,cont_trait,fam)

#Determine influential clades:
clade.test <- clade_phylm(cont_trait~pred,data=dat,phy=tree,clade.col="fam")
}
}
\author{
Gustavo Paterno
}
\references{
Here still: reference to phylolm paper + our own?
}
\seealso{
\code{\link[phylolm]{phylolm}}, \code{\link{influ_phylm}},
\code{\link{sensi_plot}}
}

