% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/influ_gls.R
\name{influ_gls}
\alias{influ_gls}
\title{Leave-one-out-deletion analysis for gls phylogenetic regression.}
\usage{
influ_gls(formula, data, phy)
}
\arguments{
\item{formula}{A model formula}

\item{phy}{A phylogeny (class 'phylo') to be matched to the data above}

\item{A}{data frame containing variables that can be attributed to the taxa
at the tips of a phylogeny}
}
\value{
The function \code{influ_gls} returns a list with the following
  components:

\code{errors} Species that showed erros during pgls fit

\code{formula} The model formula

\code{model_estimates} Full model estimates

\code{beta95_IC} Full model beta 95 confidence interval

\code{influential_species} Most influential species for beta and
  intercept

\code{results} A data frame with all simulation estimates. DFbeta and
  DFintercept represent absolute difference between full model and simulation.
}
\description{
\code{influ_gls} performs leave-one-out-deletion analysis for
\code{\link{gls}} phylogenetic linear regression. It removes one species at a time, fits
a gls model without the species and store the results of the model
estimates. The function repeats this procedure for all species in the dataset
and store all the results in a data.frame.
}
\details{
This functions only works for simple linear regression \eqn{y = bx +
  a}. Future implementation will deal with more complex models. This functions only
  works with Pagel correlation (see \code{corPagel}). If any errors occurs during simulation,
  check 'errors' to see which deletion showed error.
}
\section{Warning}{
 This code is note fully checked. Please be aware.
}
\seealso{
\code{\link{pgls}}, \code{\link{samp_pgls}}
}

