% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/influ_pgls.R
\name{influ_pgls}
\alias{influ_pgls}
\title{Leave-one-out-deletion analysis for pgls regression.}
\usage{
influ_pgls(formula, data, lambda = "ML")
}
\arguments{
\item{formula}{A model formula}

\item{data}{A 'comparative.data' object containing the covariance matrix and
data to be used in the model.}

\item{lambda}{A value for the lambda transformation. If NULL, lambda= "ML}
}
\value{
The function \code{influpgls} returns a list with the following
 components:

\code{errors} Species removal that showed erros during pgls fit

\code{formula} The model formula

\code{model_estimates} Full model estimates

\code{beta95_IC} Full model beta 95 confidence interval

\code{influential_species} Most influential species for beta and
intercept

\code{results} A data frame with all simulation estimates. DFbeta and
DFintercept represent absolute difference between full model and simulation
estimate, beta and intercept respectively.
}
\description{
\code{influ_pgls} performs leave-one-out-deletion analysis for
\code{\link{pgls}} linear regression. It removes one species at a time, fits
a pgls model without the species and store the results of the model
estimates. The function repeats this procedure for all species in the dataset
and store all the results in a data.frame.
}
\details{
This functions only works for simple linear regression
\eqn{y = bx + a}.
Future implementation will deal with more complex models.
}
\section{Warning}{
 This code is note fully checked. Please be aware.
}
\seealso{
\code{\link{[caper]pgls}}
}

