% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/influ_phyloglm.R
\name{influ_phyglm}
\alias{influ_phyglm}
\title{Influential species detection - Phylogenetic Logistic Regression}
\usage{
influ_phyglm(formula, data, phy, btol = 50, cutoff = 2, track = TRUE, ...)
}
\arguments{
\item{formula}{The model formula}

\item{data}{Data frame containing species traits with row names matching tips
in \code{phy}.}

\item{phy}{A phylogeny (class 'phylo') matching \code{data}.}

\item{btol}{Bound on searching space. For details see \code{phyloglm}}

\item{cutoff}{The cutoff value used to identify for influential species
(see Details)}

\item{track}{Print a report tracking function progress (default = TRUE)}

\item{...}{Further arguments to be passed to \code{phylolm}}
}
\value{
The function \code{influ_phyglm} returns a list with the following
components:

\code{cutoff}: The value selected for \code{cutoff}

\code{formula}: The formula

\code{full.model.estimates}: Coefficients, aic and the optimised
value of the phylogenetic parameter (i.e. \code{alpha}) for the full model
without deleted species.

\code{influential_species}: List of influential species, both
based on standardised difference in interecept and in the slope of the
regression. Species are ordered from most influential to less influential and
only include species with a standardised difference > \code{cutoff}.

\code{influ.model.estimates}: A data frame with all simulation
estimates. Each row represents a deleted clade. #' Columns report the calculated
regression intercept (\code{intercept}), difference between simulation
intercept and full model intercept (\code{DFintercept}), the standardised
difference (\code{sDFintercept}), the percentage of change in intercept compared
to the full model (\code{intercept.perc}) and intercept p-value
(\code{pval.intercept}). All these parameters are also reported for the regression
slope (\code{DFslope} etc.). Additionally, model aic value (\code{AIC}) and
the optimised value (\code{optpar}) of the phylogenetic parameter
(i.e. \code{alpha}) are reported.

\code{data}: Original full dataset.

\code{errors}: Species where deletion resulted in errors.
}
\description{
Performs leave-one-out deletion analyis for phylogenetic logistic regression,
and detects influential species.
}
\details{
This function sequentially removes one species at a time, fits a phylogenetic
logistic regression model using \code{\link[phylolm]{phyloglm}}, stores the
results and detects influential species.

Currently only logistic regression using the "logistic_MPLE"-method from
\code{phyloglm} is implemented.

\code{influ_phyglm} detects influential species based on the standardised
difference in intercept and/or slope when removing a given species compared
to the full model including all species. Species with a standardised difference
above the value of \code{cutoff} are identified as influential. The default
value for the cutoff is 2 standardised differences change.

Currently, this function can only implement simple models (i.e.
\eqn{y = a + bx}). In the future we will implement more complex models.

Output can be visualised using \code{sensi_plot}.
}
\author{
Gustavo Paterno & Gijsbert D.A. Werner
}
\references{
Here still: reference to phylolm paper + our own?
}
\seealso{
\code{\link[phylolm]{phyloglm}}, \code{\link{samp_phyglm}},
\code{\link{influ_phylm}}, \code{\link{sensi_plot}}
}

