% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/influ_phyloglm.R
\name{influ_phyloglm}
\alias{influ_phyloglm}
\title{GW working on updating help.}
\usage{
influ_phyloglm(formula, data, phy, btol = 50, cutoff = 2, ...)
}
\description{
GW working on updating help.
}

