% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samp_pgls.R
\name{samp_pgls}
\alias{samp_pgls}
\title{Sampling effort analysis for pgls linear regression.}
\usage{
samp_pgls(formula, data, phy, times = 20, breaks = seq(0.1, 0.7, 0.1))
}
\arguments{
\item{formula}{A model formula}

\item{data}{A data frame containing variables that can be attributed to the taxa
at the tips of a phylogeny}

\item{phy}{A phylogeny (class 'phylo') to be matched to the data above}

\item{times}{The number of times to repeat each simulation (per
\code{breaks}) interval.}

\item{breaks}{Percentage intervals to remove species. For example:
\code{breaks = c(.1,.2,.3)},removes 10,20 and 30 percentage of species at
random in each simulation.}
}
\value{
The function \code{samp_gls} returns a list with the following
  components:

\code{model_estimates} Full model estimates

\code{results} A data frame with all simulation estimates.

\code{power_analysis} A data frame with power analysis

\code{data} Original dataset
}
\description{
\code{samp_pgls} performs sample size sensitive analysis for \code{pgls}
linear regressions. It removes species at random, fits a pgls model without the
species and store the results of the model estimates. The percentage of
species removed is specified with \code{breaks} and the number of simulations
per break is defined by \code{times}.
}
\details{
This functions only works for simple linear regression \eqn{y = bx +
  a}. Future implementation will deal with more complex models.
}
\section{Warning}{
 This code is note fully checked. Please be aware.
}
\examples{
library(caper);library(ggplot2);library(gridExtra);library(sensiC)
data(shorebird)
comp.data <- comparative.data(shorebird.tree, shorebird.data, Species, vcv=TRUE, vcv.dim=3)
sp.ord <- match(shorebird.tree$tip.label, rownames(shorebird.data))
shorebird.data <- shorebird.data[sp.ord,]
samp1 <- samp_pgls(log(Egg.Mass) ~ log(M.Mass),data=shorebird.data,phy=shorebird.tree)
sensi_plot(samp1)
}
\seealso{
\code{\link{pgls}}, \code{\link{influ_gls}}
}

