% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samp_phyloglm.R
\name{samp_phyglm}
\alias{samp_phyglm}
\title{Sensitivity Analysis Species Sampling  - Phylogenetic Logistic Regression}
\usage{
samp_phyglm(formula, data, phy, times = 20, breaks = seq(0.1, 0.7, 0.1),
  btol = 50, track = TRUE, ...)
}
\arguments{
\item{formula}{The model formula}

\item{data}{Data frame containing species traits with row names matching tips
in \code{phy}.}

\item{phy}{A phylogeny (class 'phylo') matching \code{data}.}

\item{times}{The number of times species are randomly deleted for each
\code{break}.}

\item{breaks}{A vector containing the percentages of species to remove.}

\item{btol}{Bound on searching space. For details see \code{phyloglm}}

\item{track}{Print a report tracking function progress (default = TRUE)}

\item{...}{Further arguments to be passed to \code{phylolm}}
}
\value{
The function \code{samp_phylolm} returns a list with the following
components:

\code{formula}: The formula

\code{full.model.estimates}: Coefficients, aic and the optimised
value of the phylogenetic parameter (e.g. \code{lambda} or \code{kappa}) for
the full model without deleted species.

\code{samp.model.estimates}: A data frame with all simulation
estimates. Each row represents a model rerun with a given number of species
\code{n.remov} removed, representing \code{n.percent} of the full dataset.
Reported are the calculated regression intercept (\code{intercept}),
difference between simulation intercept and full model intercept (\code{DFintercept}),
the percentage change in intercept compared to the full model (\code{intercept.perc})
and intercept p-value (\code{pval.intercept}). All of these are also reported
for the regression slope (\code{DFslope} etc.). Additonally, model aic value
(\code{AIC}) and the optimised value (\code{optpar}) of the phylogenetic
parameter (e.g. \code{kappa} or \code{lambda}, depends on phylogeneticmodel
used) are reported.

\code{sign.analysis} For each break (i.e. each percentage of species
removed) this reports the percentage of statistically signficant (at p<0.05)
intercepts (\code{perc.sign.intercept}) over all repititions as well as the
percentage of statisticaly significant (at p<0.05) slopes (\code{perc.sign.slope}).

\code{data}: Original full dataset.
}
\description{
Performs analyses of sensitivity to species sampling by randomly removing
species and detecting the effects on parameter estimates in phylogenetic
logistic regression.
}
\details{
This function randomly removes a given percentage of species (controlled by
\code{breaks}) from the full phylogenetic logistic regression, fits a phylogenetic
logistic regression model without these species using \code{\link[phylolm]{phyloglm}},
repeates this many times (controlled by \code{times}), stores the results and
calculates the effects on model parameters.

Only logistic regression using the "logistic_MPLE"-method from
\code{phyloglm} is implemented.

Currently, this function can only implement simple linear models (i.e. \eqn{trait~
predictor}). In the future we will implement more complex models.

Output can be visualised using \code{sensi_plot}.
}
\examples{
\dontrun{
library(sensiPhy);library(phylolm)

#Generate a random tree
set.seed(2468)
tree <- rtree(100)

#Generate random predictor variable (pred), evolving according to a BM model.
pred<- rTraitCont(tree,root.value=0,sigma=1,model="BM")

#Generate two continous traits, one evolving highly correlated with the
#predictor (trait 1), and one evolving more randomly (trait 2)
cont_trait1 <- pred + rTraitCont(tree,model="BM",sigma=0.1)
cont_trait2 <- pred + rTraitCont(tree,model="BM",sigma=10)

#Generate two binary traits, one highly correlated to pred (trait 1), the other less.
bin_trait1<-rbinTrait(n=1,tree,beta=c(-1,0.5),alpha=0.1,
                     X=cbind(rep(1,length(tree$tip.label)),pred))
bin_trait2<-rbinTrait(n=1,tree,beta=c(-1,0.5),alpha=5,
                      X=cbind(rep(1,length(tree$tip.label)),pred))
dat<-data.frame(pred,cont_trait1,cont_trait2,bin_trait1,bin_trait2)

#For both regressions, determine sensitivity of results to species sampled.
fit1<-samp_phyglm(bin_trait1~pred,data = dat,phy = tree)
fit2<-samp_phyglm(bin_trait2~pred,data = dat,phy = tree)

#It is possible to change the species removal percentages and number of repeats
fit3<-samp_phyglm(bin_trait2~pred,data = dat,phy = tree,
     breaks = c(0.25,0.5,0.75),times=100)
}
}
\author{
Gustavo Paterno & Gijsbert D.A. Werner
}
\references{
Here still: reference to phylolm paper + our own + some more
background papers.
}
\seealso{
\code{\link[phylolm]{phyloglm}}, \code{\link{samp_phylolm}},
\code{\link{influ_phyloglm}}, \code{\link{sensi_plot}}
}

