% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sensi_plot.R
\name{sensi_plot}
\alias{sensi_plot}
\title{Graphical sensitive analysis for comparative methods}
\usage{
sensi_plot(x, graphs = "all", param = "slope")
}
\arguments{
\item{x}{output from \code{samp_gls}, \code{influ_gls}}

\item{graphs}{choose which graphs should be printed on the output ("all", 1,2,3 or 4)}
}
\description{
\code{sensi_plot} Plot results from \code{samp_influ_phyloglm},
\code{influ_pgls}
}

