% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sensi_plot.R
\name{sensi_plot}
\alias{sensi_plot}
\title{Plot results from package `sensiC`}
\usage{
sensi_plot(x)
}
\arguments{
\item{x}{output from \code{samp_pgls}, \code{influ_pgls}, \code{samp_gls} and
code{influ_gls}}
}
\description{
\code{sensi_plot} Plot results from \code{samp_pgls}, \code{samp_gls},
\code{influ_pgls},\code{influ_gls}
}

