% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sensi_plot.R
\name{sensi_plot}
\alias{sensi_plot}
\title{Plot results from package `sensiC`}
\usage{
sensi_plot(x, method = "sampling")
}
\description{
\code{sensi_plot} Performs Sampling effort analysis for \code{pgls}
regression. It removes species at random, fits a pgls model without the
species and store the results of the model estimates. The percentage of
species removed is specified with \code{breaks} and the number of simulations
per break is defined by \code{times}.
}

