% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_influ_phylolm.R
\name{sensi_plot.sensiInflu}
\alias{sensi_plot.sensiInflu}
\title{Graphical diagnostics for class 'sensiInflu'}
\usage{
\method{sensi_plot}{sensiInflu}(x, graphs = "all", param = "slope", ...)
}
\arguments{
\item{x}{output from \code{influ_phylolm}}

\item{graphs}{choose which graph should be printed on the output ("all", 1,2,3 or 4)}

\item{param}{choose which parameter ("intercept" or "slope" should be printed)}

\item{...}{further arguments to methods}
}
\description{
\code{plot_influ_phylolm} Plot results from \code{influ_phylolm} and
\code{influ_phyloglm}
}
\details{
For 'x' from influ_phylolm or influ_phyloglm:

Graph 1: Distribution of estimated slopes or intercepts for each
simulation (leave-one-out deletion). Red vertical line represents the original
slope or intercept from the full model (with all species).

Graph 2: Original regression plot (\eqn{trait~predictor}). Standardized
difference in slope or intercept is represented by a continous colour scale,
ranging from black (low \code{sDFintercept} or \code{sDFslope} values) to red
(high \code{sDFintercept} or \code{sDFslope} values).

Graph 3: Distribution of Standardized difference in slope or intercept. Red
colour indicates inbfluential species (with a standardised difference above
the value of \code{cutoff}).

Graph 4: Ditribution of the percentage of change in slope or intercept.
}
\examples{
\dontrun{
library(sensiPhy)

#Generate a random tree
set.seed(2468)
tree <- rtree(100)

#Generate random predictor variable (pred), evolving according to a BM model.
pred<- rTraitCont(tree,root.value=0,sigma=1,model="BM")

#Generate two continous traits, one evolving highly correlated with the
#predictor (trait 1), and one evolving more randomly (trait 2)
cont_trait <- pred + rTraitCont(tree,model="BM",sigma=0.1)

#Generate two binary traits, one highly correlated to pred (trait 1), the other less.
bin_trait <-rbinTrait(n=1,tree,beta=c(-1,0.5),alpha=0.1,
                     X=cbind(rep(1,length(tree$tip.label)),pred))
dat<-data.frame(pred,cont_trait,bin_trait)

#Determine influential species for both regressions.
fit1<-influ_phylolm(cont_trait~pred,data = dat,phy = tree)
fit2<-influ_phyloglm(bin_trait~pred,data = dat,phy = tree)

# Plot results:
sensi_plot(fit1)
sensi_plot(fit2)
# You can also choose which graph and parameter should be ploted:
sensi_plot(fit1, graphs = 1, param = "intercept")
sensi_plot(fit2, graphs = "all", param = "slope")
}
}
\author{
Gustavo Paterno
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}

