% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi_plot.sensiINTER_Samp.R
\name{sensi_plot.sensiIntra_Samp}
\alias{sensi_plot.sensiIntra_Samp}
\title{Graphical diagnostics for class 'sensiIntra_Samp' and 'sensiTree_samp'}
\usage{
\method{sensi_plot}{sensiIntra_Samp}(x, graphs = "all", param = "estimate",
  ...)
}
\arguments{
\item{x}{output from \code{sensiIntra_Samp_phylm} or \code{sensiIntra_Samp_phyglm}}

\item{graphs}{choose which graph should be printed on the output ("all", 1,2,3 or 4)}

\item{param}{choose which model parameter should be ploted  ("intercept" or "estimate")}

\item{...}{further arguments to methods}
}
\description{
\code{sensi_plot.sensiIntra_Samp} Plot results from \code{sensiIntra_Samp_phylm} and
\code{sensiIntra_Samp_phyglm}
}
\details{
For 'x' from sensiIntra_Samp_phylm or sensiIntra_Samp_phyglm:

Graph 1: Estimated estimates (slopes) or intercepts for each simution across  
percentages of species removed and intra simulations. Colours represent percentage 
of change in comparison with the full model (blue = lower than 5, orange = 
between 5 and 10 and red = higher than 10).
The red horizontal line represents the original slope or 
intercept from the full model (with all species). 

Graph 2: The proportion of estimated slopes and intercepts in each category 
across the percentage of species removed.

Graph 3: Estimated phylogenetic model parameter for each simulation across
the percentage of species removed.

Graph 4: The percentage of significant slopes or intercepts across the 
percentage of species removed.
}
\note{
If model = "BM", only plots 1, 2 and 4 are printed. Plot 3, phylogenetic
 model parameter is not available for model = "BM"
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[sensiPhy]{samp_phylm}}
\code{\link[sensiPhy]{samp_phyglm}}
}
\author{
Gustavo Paterno
}
