% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensi_plot.sensiINTER_Influ.R
\name{sensi_plot.sensiTree_Influ}
\alias{sensi_plot.sensiTree_Influ}
\title{Graphical diagnostics for class 'sensiTree_Influ'}
\usage{
\method{sensi_plot}{sensiTree_Influ}(x, graphs = "all", param = "estimate",
  ...)
}
\arguments{
\item{x}{output from \code{tree_influ_phylm} or \code{tree_influ_phyglm}}

\item{graphs}{choose which graph should be printed on the output ("all", 1,2,3 or 4)}

\item{param}{choose which parameter ("intercept" or "estimate" should be printed)}

\item{...}{further arguments to methods}
}
\description{
\code{sensi_plot.sensiTree_Influ} Plot results from \code{tree_influ_phylm} and 
\code{tree_influ_phyglm}
}
\details{
For 'x' from sensiTree_Intra_phylm or sensiTree_Intra_phyglm:

Graph 1: Distribution of estimated estimates (slopes) or intercepts for each 
simulation (leave-one-out deletion). Red vertical line represents the original
slope or intercept from the full model (with all species). 

Graph 2: Original regression plot (\eqn{trait~predictor}). Standardized 
difference in slope or intercept is represented by a continous size scale. 
The names of the most influential species (sDF > cutoff) are ploted in the
graph. 

Graph 3: Distribution of standardized difference in slope or intercept. Red 
colour indicates inbfluential species (with a standardised difference above 
the value of \code{cutoff}).

Graph 4: Distribution of the percentage of change in slope or intercept.
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
\author{
Gustavo Paterno, Caterina Penone
}
