% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensiPhy.R
\docType{package}
\name{sensiPhy}
\alias{sensiPhy}
\alias{sensiPhy-package}
\title{Sensitivity Analysis for Comparative Methods}
\description{
An implementation of sensitivity analysis for phylogenetic comparative
methods. The package is an umbrella of statistical and graphical methods that 
estimate and report different types of uncertainty in PCM:  
(i) Species Sampling uncertainty (sample size; influential species and clades).  
(ii) Phylogenetic uncertainty (different topologies and/or branch lengths).  
(iii) Data uncertainty (intraspecific variation and measurement error).
}
\section{Bug reports}{
 

\url{https://github.com/paternogbc/sensiPhy/issues}.
}

\section{How to contribute?}{
 

You can find instructions on how to contribute to \bold{sensiPhy} at
this link: \url{https://github.com/paternogbc/sensiPhy/wiki/How-to-support-sensiPhy}

A quick tutorial is available at this link: 
\url{https://github.com/paternogbc/sensiPhy/wiki}
}

\author{
Gustavo Paterno; Caterina Penone & Gijsbert D.A. Werner
}
